@@ -5,20 +5,40 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.8</version>
+		<version>3.2.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>gov.epa.ccte.api</groupId>
 	<artifactId>hazard</artifactId>
-	<version>1</version>
+	<version>1.0.0</version>
 	<name>hazard</name>
 	<description>project for hazard data</description>
 	<properties>
-		<java.version>1.8</java.version>
-		<lombok.version>1.18.24</lombok.version>
+		<java.version>17</java.version>
+		<testcontainers.version>1.17.6</testcontainers.version>
+		<lombok.version>1.18.30</lombok.version>
 		<mapstruct.version>1.5.3.Final</mapstruct.version>
 	</properties>
+
+	<distributionManagement>
+        	<repository>
+            		<id>github</id>
+            		<name>GitHub OWNER Apache Maven Packages</name>
+            		<url>https://maven.pkg.github.com/USEPA/ccte-api-hazard</url>
+        	</repository>
+    	</distributionManagement>
+	
 	<dependencies>
+
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-actuator</artifactId>
+		</dependency>
+
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-web</artifactId>
+		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-data-rest</artifactId>
