@@ -1,26 +1,23 @@
 package gov.epa.ccte.api.hazard.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import gov.epa.ccte.api.hazard.domain.ApiKey;
-import gov.epa.ccte.api.hazard.repository.ApiKeyRepository;
 import gov.epa.ccte.api.hazard.web.rest.error.AuthorizationProblem;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.GenericFilterBean;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
-import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -29,76 +26,169 @@ import java.util.concurrent.ConcurrentHashMap;
 @Component
 public class ApiKeyRequestFilter extends GenericFilterBean {
 
-    private ConcurrentHashMap<UUID, String> keyStore;// = new ConcurrentHashMap();
-
+    private final ConcurrentHashMap<UUID, String> keyStore;// = new ConcurrentHashMap();
+    private final ConcurrentHashMap<String, String> approvedOriginStore;// = new ConcurrentHashMap();
     private final ObjectMapper mapper = new ObjectMapper();
     private final String keyName;
 
-    public ApiKeyRequestFilter(ApiKeyRepository repository, @Value("${application.api-key-name}") String keyName) {
+    public ApiKeyRequestFilter(ConcurrentHashMap<UUID, String> keyStore,
+                               ConcurrentHashMap<String, String> approvedOriginStore,
+                               @Value("${application.api-key-name}") String keyName) {
+        this.keyStore = keyStore;
+        this.approvedOriginStore = approvedOriginStore;
         this.keyName = keyName;
 
-        initializeKeyStore(repository);
+        //initializeKeyStore(repository);
+        //initializeApprovedOriginStore();
     }
 
-    private void initializeKeyStore(ApiKeyRepository repository) {
-        keyStore = new ConcurrentHashMap<>();
+//    private void initializeApprovedOriginStore() {
+//        approvedOriginStore = new ConcurrentHashMap<>();
+//        approvedOriginStore.put("http://localhost:3003", "http://localhost:3003");
+//        approvedOriginStore.put("http://localhost:8888", "http://localhost:8888");
+//        approvedOriginStore.put("https://ccte-ccd-dev.epa.gov", "https://ccte-ccd-dev.epa.gov");
+//        approvedOriginStore.put("https://ccte-ccd-stg.epa.gov", "https://ccte-ccd-stg.epa.gov");
+//        approvedOriginStore.put("https://ccte-ccd-prod.epa.gov", "https://ccte-ccd-prod.epa.gov");
+//        approvedOriginStore.put("https://comptox.epa.gov", "https://comptox.epa.gov");
+//        approvedOriginStore.put("https://ccte-api-s.app.cloud.gov", "https://ccte-api-s.app.cloud.gov");
+//    }
 
-        List<ApiKey> keys = repository.findAll();
-
-        for(ApiKey key : keys)
-            keyStore.put(key.getId(), key.getEmail());
-
-        log.info("*** {} keys are loaded. *** ", keys.size());
-    }
+//    private void initializeKeyStore(ApiKeyRepository repository) {
+//        keyStore = new ConcurrentHashMap<>();
+//
+//        List<ApiKey> keys = repository.findAll();
+//
+//        for(ApiKey key : keys)
+//            keyStore.put(key.getId(), key.getEmail());
+//
+//        log.info("*** {} keys are loaded. *** ", keys.size());
+//    }
 
     @Override
     public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
+
+        log.info("*** checking the API key ***");
+
+        // dump the request headers
+        log.info("*** request headers ***");
+//        Enumeration<String> headerNames = ((HttpServletRequest) servletRequest).getHeaderNames();
+//
+//        if (headerNames != null) {
+//            while (headerNames.hasMoreElements()) {
+//                String header = headerNames.nextElement();
+//                log.info("Header: {} = {}", header, ((HttpServletRequest) servletRequest).getHeader(header));
+//            }
+//        }
+
         HttpServletRequest req = (HttpServletRequest) servletRequest;
-        String path = req.getRequestURI();
 
 //        if(path.startsWith("/api") == false){
 //            filterChain.doFilter(servletRequest, servletResponse);
 //            return;
 //        }
 
-        // Get apikey from the http header
-        String key = req.getHeader(keyName) == null ? "" : req.getHeader(keyName);
-        log.debug("Trying {}: {}", keyName, key );
+        //
+        if(shouldCheckApiKey(req)) {
+            log.info("*** API key check is checking ***");
+            //String headers = String.valueOf(req.getHeaderNames());
+            String key = getApiKeyfromHttpHeader(req.getHeader(keyName));
 
-        // In case user is provided api key through requrest parameter api-key
-        if(key == null || key.equals("")){
-            String queryString = req.getQueryString();
-            if(StringUtils.isNoneEmpty(queryString)){
-                queryString = URLDecoder.decode(queryString, StandardCharsets.UTF_8.toString());
-                int pos = StringUtils.indexOfIgnoreCase(queryString, keyName + "=");
-                String keyValue = StringUtils.substring(queryString, pos+8);
-                key = StringUtils.substringBefore(keyValue,"&");
-                // key = StringUtils.substringBetween(queryString, "api_key=","&");
+            // In case user is provided api key through parameter x-api-key
+            if (key == null || key.isEmpty()) {
+                // get key from the URL parameter
+                key = getApiKeyFromQueryParam(req.getQueryString());
+            }
+
+            if (key == null || key.isEmpty()) {
+                // api key is missing
+                returnErrorMsg(servletResponse, key);
+            } else if (isKeyExist(key)) {
+                // api key found
+                filterChain.doFilter(servletRequest, servletResponse);
+            } else {
+                // api doesn't match
+                returnErrorMsg(servletResponse, key);
+                log.info("*** API key {} not recognized *** ", key);
+            }
+        }else{
+            log.info("*** API key check is skipped ***");
+            filterChain.doFilter(servletRequest, servletResponse);
+        }
+    }
+
+    private boolean shouldCheckApiKey(HttpServletRequest req) {
+        // check if request has method OPTIONS or remote is localhost
+        String method = req.getMethod();    // OPTIONS
+        String origin = Optional.ofNullable(req.getHeader("origin")).orElse("");  // example - origin = http://localhost:8888
+        String referer = Optional.ofNullable(req.getHeader("Referer")).orElse("");  // example - referer = http://localhost:8888/dashboard
+        String path = Optional.ofNullable(req.getServletPath()).orElse(""); // example - path = /chemical/file/image/search/by-dtxsid/DTXSID7020182
+
+        String refererdHost;
+        if(!referer.isEmpty()){
+            refererdHost = "https://" + referer.split("/")[2];  // example - referredHost = localhost:8888
+        }else{
+            refererdHost = ""; //"https://" + req.getHeader("Referer").split("/")[2];  // example - referredHost = localhost:8888
+        }
+
+        log.debug("method = {}, origin = {}, referer ={}, refererdHost = {}, path={} ",method, origin, referer, refererdHost, path);
+
+        // not security for health endpoint
+        if( path.contains("/hazard/health")){
+            log.debug("skipping api-key check");
+            return false;
+        }
+
+//        if(method.equalsIgnoreCase("OPTIONS") || approvedOrigin(origin) || approvedOrigin(refererdHost))
+        if(approvedOrigin(origin) || approvedOrigin(refererdHost))
+            return false;
+        else
+            return true;
+    }
+
+    private boolean approvedOrigin(String origin) {
+
+        return origin != null && approvedOriginStore.containsKey(origin);
+    }
+
+
+    private String getApiKeyFromQueryParam(String query) {
+
+        // an example -  format=svg&x-api-key=f1d96bdd-223a-434e-b1c0-af373a59a19e
+
+        if(query != null){
+            String[] params = query.split("&");
+
+            for(String param: params){
+                int idx = param.indexOf("=");
+                if(param.substring(0,idx).equalsIgnoreCase(keyName)){
+                    return URLDecoder.decode(param.substring(idx + 1), StandardCharsets.UTF_8);
+                }
             }
         }
+        return "";
+    }
 
-        if(key == null || key.equals("")){
-            returnErrorMsg(servletResponse, key);
-        } else if(isKeyExist(key)){
-            filterChain.doFilter(servletRequest, servletResponse);
-        }else{
-            returnErrorMsg(servletResponse, key);
+    private String getApiKeyfromHttpHeader(String key) {
+        // Get apikey from the http header
+        //String key = req.getHeader(keyName) == null ? "" : req.getHeader(keyName);
 
-            log.info("*** API key {} not recognized *** ", key);
+        if(key == null || key.isEmpty()){
+            log.debug("Custom http  header {} not found", keyName);
         }
+
+        return key;
     }
 
     private void returnErrorMsg(ServletResponse servletResponse, String key) throws IOException {
         HttpServletResponse resp = (HttpServletResponse) servletResponse;
-        String error = "Invalid API KEY";
 
         AuthorizationProblem problem;
 
-        if(key == null || key.equals("")){
+        if(key == null || key.isEmpty()){
             // header is missing
             problem = AuthorizationProblem.builder()
                     .title("API Header Not Found")
-                    .detail("Every API call should pass assigned API key through http header. Request is missing header " + keyName+ "." )
+                    .detail("Every API call should pass assigned API key through custom http header or query parameter. Request is missing " + keyName+ "." )
                     .build();
 
         }else{
